@@ -3,6 +3,7 @@
     "description": "the new thinkphp framework",
     "type": "yzncms-framework",
     "keywords": [
+        "yzncms",
         "framework",
         "thinkphp",
         "ORM"
@@ -20,8 +21,9 @@
         }
     ],
     "require": {
-        "php": ">=5.6.0",
-        "yzncms/think-installer": "2.*"
+        "php": ">=7.0.0",
+        "yzncms/think-installer": "2.*",
+        "psr/http-message": "^1.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^5.0|^6.0",
