@@ -57,7 +57,16 @@ class Alioss extends Addons
         $endpoint = $config['endpoint'];
         // 存储空间名称
         $bucket = $config['bucket'];
+        //图片处理的规则名称stylename
+        $stylename = '';
+        if (isset($config['stylename']) && $config['stylename']) {
+            if (isset($config['separator']) && $config['separator']) {
+                $stylename = trim($config['separator']) . trim($config['stylename']);
+            } else {
+                $stylename = '?x-oss-process=style/' . trim($config['stylename']);
+            }
 
+        }
         // 文件信息
         $info   = $file->getInfo();
         $suffix = strtolower(pathinfo($info['name'], PATHINFO_EXTENSION));
@@ -87,7 +96,7 @@ class Alioss extends Addons
             'uid'    => (int) cookie('uid'),
             'name'   => $info['name'],
             'mime'   => $info['type'],
-            'path'   => $config['domain'] . $object,
+            'path'   => $config['domain'] . $object . $stylename,
             'ext'    => $suffix,
             'size'   => $file->getSize(),
             'md5'    => $file->hash('md5'),
