@@ -1,31 +1,45 @@
 <?php
 return [
-    0 => [
+    [
         'name'  => 'accessKey',
         'title' => '七牛云的ak',
         'type'  => 'text',
         'value' => '',
         'tip'   => '七牛云的ak',
     ],
-    1 => [
+    [
         'name'  => 'secrectKey',
         'title' => '七牛云的sk',
         'type'  => 'text',
         'value' => '',
         'tip'   => '七牛云的sk',
     ],
-    2 => [
+    [
         'name'  => 'bucket',
         'title' => '七牛云的空间名称',
         'type'  => 'text',
         'value' => '',
         'tip'   => '七牛云的空间名称',
     ],
-    3 => [
+    [
         'name'  => 'domain',
         'title' => '七牛云绑定的域名',
         'type'  => 'text',
         'value' => '',
         'tip'   => '七牛云的空间对应的域名，已http://开头，/结尾',
+    ],
+    [
+        'name'  => 'stylename',
+        'title' => '图片样式名称',
+        'type'  => 'text',
+        'value' => '',
+        'tip'   => '设置后，同时在阿里云oss后台【访问设置】开启原图保护',
+    ],
+    [
+        'name'  => 'separator',
+        'title' => '图片样式分隔符',
+        'type'  => 'text',
+        'value' => '',
+        'tip'   => '配合图片样式，同时在七牛云后台设置分隔符',
     ],
 ];
