@@ -29,7 +29,7 @@ class Adminaddon extends Adminbase
     {
 
         parent::initialize();
-        $this->addonName = \think\Loader::parseName($this->request->controller());
+        $this->addonName = parse_name($this->request->controller());
         $this->addonPath = ADDON_PATH . $this->addonName . DIRECTORY_SEPARATOR;
     }
 
