@@ -15,6 +15,6 @@
 
 return [
     "yzncms_name"    => "YznCMS内容管理系统", //产品名称
-    "yzncms_release" => 20220321, //产品更新日期
+    "yzncms_release" => 20220606, //产品更新日期
     "yzncms_version" => "1.0.0", //产品版本号
 ];
